requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
pydantic
reportlab
PyPDF2
Pillow
